backend/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
geopandas==0.14.2
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1
